+++ runway/blueprints/staticsite/dependencies.py
@@ -9,6 +9,8 @@ from awacs.aws import Allow, AWSPrincipal, Policy, Statement
 from troposphere import AccountId, Join, Output, cognito, s3
 
 from runway.cfngin.blueprints.base import Blueprint
+from runway.hooks.staticsite.auth_at_edge.client_updater import get_redirect_uris
+from runway.module.staticsite import add_url_scheme
 
 LOGGER = logging.getLogger(__name__)
 
@@ -17,6 +19,17 @@ class Dependencies(Blueprint):
     """Stacker blueprint for creating static website buckets."""
 
     VARIABLES = {
+        "Aliases": {
+            "type": list,
+            "default": [],
+            "description": "(Optional) Domain aliases for the distribution",
+        },
+        "AdditionalRedirectDomains": {
+            "type": list,
+            "default": [],
+            "description": "(Optional) AppClient callback/logout domains (in "
+            "addition to the Aliases)",
+        },
         "AuthAtEdge": {
             "type": bool,
             "default": False,
@@ -26,6 +39,10 @@ class Dependencies(Blueprint):
             "type": bool,
             "default": False,
             "description": "Whether a User Pool should be created for the project",
+        },
+        "SupportedIdentityProviders": {
+            "type": list,
+            "description": "Auth@Edge AppClient Identity Providers",
         },
         "OAuthScopes": {
             "type": list,
@@ -37,6 +54,16 @@ class Dependencies(Blueprint):
                 "aws.cognito.signin.user.admin",
             ],
             "description": "The allowed scopes for OAuth validation",
+        },
+        "RedirectPathSignIn": {
+            "type": str,
+            "default": "/parseauth",
+            "description": "Auth@Edge redirect sign in path",
+        },
+        "RedirectPathSignOut": {
+            "type": str,
+            "default": "/",
+            "description": "Auth@Edge redirect sign out path",
         },
     }
 
@@ -107,9 +134,30 @@ class Dependencies(Blueprint):
         )
 
         if variables["AuthAtEdge"]:
-            callbacks = self.context.hook_data["aae_callback_url_retriever"][
-                "callback_urls"
-            ]
+            userpool_client_params = {
+                "AllowedOAuthFlows": ["code"],
+                "AllowedOAuthScopes": variables["OAuthScopes"],
+            }
+            if variables["Aliases"]:
+                userpool_client_params["AllowedOAuthFlowsUserPoolClient"] = True
+                userpool_client_params["SupportedIdentityProviders"] = variables[
+                    "SupportedIdentityProviders"
+                ]
+
+                redirect_domains = [add_url_scheme(x) for x in variables["Aliases"]] + [
+                    add_url_scheme(x) for x in variables["AdditionalRedirectDomains"]
+                ]
+                redirect_uris = get_redirect_uris(
+                    redirect_domains,
+                    variables["RedirectPathSignIn"],
+                    variables["RedirectPathSignOut"],
+                )
+                userpool_client_params["CallbackURLs"] = redirect_uris["sign_in"]
+                userpool_client_params["LogoutURLs"] = redirect_uris["sign_out"]
+            else:
+                userpool_client_params["CallbackURLs"] = self.context.hook_data[
+                    "aae_callback_url_retriever"
+                ]["callback_urls"]
 
             if variables["CreateUserPool"]:
                 user_pool = template.add_resource(
@@ -129,15 +177,10 @@ class Dependencies(Blueprint):
                 user_pool_id = self.context.hook_data["aae_user_pool_id_retriever"][
                     "id"
                 ]
+            userpool_client_params["UserPoolId"] = user_pool_id
 
             client = template.add_resource(
-                cognito.UserPoolClient(
-                    "AuthAtEdgeClient",
-                    AllowedOAuthFlows=["code"],
-                    CallbackURLs=callbacks,
-                    UserPoolId=user_pool_id,
-                    AllowedOAuthScopes=variables["OAuthScopes"],
-                )
+                cognito.UserPoolClient("AuthAtEdgeClient", **userpool_client_params)
             )
 
             template.add_output(